requirements.txt
fastapi
uvicorn
python-dotenv
python-supabase
jinja2
